fix: 금액 소수점 제거 - 모든 가격/수익 필드를 정수로 변환
- trade_store: DB 조회 시 entry_price, exit_price, pnl → int()
- position_manager: total_equity, cash, daily_pnl → int()
- trading_engine: 포지션/거래 가격 필드 → int()
- risk_manager: daily_pnl, daily_loss_limit, total_equity → int()
- analysis routes: round(x, 0) → int(round(x))
- pnl_pct, win_rate 등 비율만 소수점 유지

+++ src/engine/risk_manager.py
@@ -135,7 +135,7 @@ class RiskManager:
             "circuit_breaker": self._circuit_breaker_active,
             "positions": self.pm.position_count,
             "max_positions": self.max_positions,
-            "daily_pnl": self.pm.daily_pnl,
-            "daily_loss_limit": -self.pm.initial_capital * self.max_daily_loss_pct,
-            "total_equity": self.pm.total_equity,
+            "daily_pnl": int(self.pm.daily_pnl),
+            "daily_loss_limit": int(-self.pm.initial_capital * self.max_daily_loss_pct),
+            "total_equity": int(self.pm.total_equity),
         }
